# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AEROVAULTT Attendance API specification
servers:
  - url: /api
    description: Base API path
...
security:
  - bearerAuth: []
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List employees (HR/Admin)
      parameters:
        - name: departmentId
          in: query
          schema: { type: integer }
        - name: isActive
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create a new employee (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get an employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update an employee (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}/reset-password:
    post:
      operationId: resetEmployeePassword
      tags: [employees]
      summary: Reset an employee's password (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /employees/{id}/deactivate:
    post:
      operationId: deactivateEmployee
      tags: [employees]
      summary: Deactivate or reactivate an employee (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeactivateInput"
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
...
  /office-locations:
    get:
      operationId: listOfficeLocations
      tags: [officeLocations]
      summary: List office locations
      responses:
        "200":
          description: List of office locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OfficeLocation"
    post:
      operationId: createOfficeLocation
      tags: [officeLocations]
      summary: Create an office location (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfficeLocationInput"
      responses:
        "201":
          description: Office location created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfficeLocation"
  /office-locations/{id}:
    patch:
      operationId: updateOfficeLocation
      tags: [officeLocations]
      summary: Update an office location (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfficeLocationInput"
      responses:
        "200":
          description: Updated office location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfficeLocation"
    delete:
      operationId: deleteOfficeLocation
      tags: [officeLocations]
      summary: Delete an office location (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
  /leaves:
    get:
      operationId: listLeaves
      tags: [leaves]
      summary: List leave requests (own for employees, all for HR/Admin)
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/LeaveStatus"
        - name: employeeId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Leave requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Leave"
    post:
      operationId: createLeave
      tags: [leaves]
      summary: Submit a leave request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveInput"
      responses:
        "201":
          description: Leave request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Leave"
  /leaves/{id}:
    patch:
      operationId: reviewLeave
      tags: [leaves]
      summary: Approve or reject a leave request (HR/Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveReview"
      responses:
        "200":
          description: Updated leave request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Leave"
  /holidays:
    get:
      operationId: listHolidays
      tags: [holidays]
      summary: List holidays
      responses:
        "200":
          description: Holidays
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Holiday"
    post:
      operationId: createHoliday
      tags: [holidays]
      summary: Create a holiday (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HolidayInput"
      responses:
        "201":
          description: Holiday created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Holiday"
  /holidays/{id}:
    delete:
      operationId: deleteHoliday
      tags: [holidays]
      summary: Delete a holiday (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }
    Role:
      type: string
      enum: [employee, hr, admin]
...
    LeaveType:
      type: string
      enum: [casual, sick, permission, wfh]
    LeaveStatus:
      type: string
      enum: [pending, approved, rejected]
...
    ResetPasswordInput:
      type: object
      required: [newPassword]
      properties:
        newPassword: { type: string, minLength: 6 }
    DeactivateInput:
      type: object
      required: [isActive]
      properties:
        isActive: { type: boolean }
    Employee:
      type: object
      required:
        - id
        - employeeId
        - name
        - email
        - role
        - isActive
        - createdAt
      properties:
        id: { type: integer }
        employeeId: { type: string }
        name: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        role:
          $ref: "#/components/schemas/Role"
        departmentId: { type: ["integer", "null"] }
        departmentName: { type: ["string", "null"] }
        officeLocationId: { type: ["integer", "null"] }
        officeLocationName: { type: ["string", "null"] }
        shiftStart: { type: ["string", "null"] }
        shiftEnd: { type: ["string", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
    EmployeeInput:
      type: object
      required: [employeeId, name, email, password, role]
      properties:
        employeeId: { type: string }
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        password: { type: string, minLength: 6 }
        role:
          $ref: "#/components/schemas/Role"
        departmentId: { type: integer }
        officeLocationId: { type: integer }
        shiftStart: { type: string }
        shiftEnd: { type: string }
    EmployeeUpdate:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        role:
          $ref: "#/components/schemas/Role"
        departmentId: { type: integer }
        officeLocationId: { type: integer }
        shiftStart: { type: string }
        shiftEnd: { type: string }
...
    OfficeLocation:
      type: object
      required: [id, name, latitude, longitude, radiusMeters]
      properties:
        id: { type: integer }
        name: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        radiusMeters: { type: integer }
    OfficeLocationInput:
      type: object
      required: [name, latitude, longitude, radiusMeters]
      properties:
        name: { type: string, minLength: 1 }
        latitude: { type: number }
        longitude: { type: number }
        radiusMeters: { type: integer, minimum: 1 }
...
    Leave:
      type: object
      required:
        - id
        - employeeId
        - employeeName
        - type
        - startDate
        - endDate
        - reason
        - status
        - createdAt
      properties:
        id: { type: integer }
        employeeId: { type: integer }
        employeeName: { type: string }
        type:
          $ref: "#/components/schemas/LeaveType"
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        reason: { type: string }
        attachmentUrl: { type: ["string", "null"] }
        status:
          $ref: "#/components/schemas/LeaveStatus"
        reviewNotes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    LeaveInput:
      type: object
      required: [type, startDate, endDate, reason]
      properties:
        type:
          $ref: "#/components/schemas/LeaveType"
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        reason: { type: string, minLength: 1 }
        attachmentUrl: { type: string }
    LeaveReview:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/LeaveStatus"
        reviewNotes: { type: string }
    Holiday:
      type: object
      required: [id, name, date]
      properties:
        id: { type: integer }
        name: { type: string }
        date: { type: string, format: date }
    HolidayInput:
      type: object
      required: [name, date]
      properties:
        name: { type: string, minLength: 1 }
        date: { type: string, format: date }
    Notification:
      type: object
      required: [id, message, type, isRead, createdAt]
      properties:
        id: { type: integer }
        message: { type: string }
        type: { type: string }
        isRead: { type: boolean }
        createdAt: { type: string, format: date-time }
...