# lib/api-zod/src/generated/types/leaveType.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AEROVAULTT Attendance API specification
 * OpenAPI spec version: 0.1.0
 */

export type LeaveType = typeof LeaveType[keyof typeof LeaveType];


export const LeaveType = {
  casual: 'casual',
  sick: 'sick',
  permission: 'permission',
  wfh: 'wfh',
} as const;